Add Library page and update navigation
